misseditor: fixed mission editor to work under MacOS
